modfs: add script for volatile directory on NAS storage
If you're using a NAS directory to store files at a higher frequency
(e.g. pictures from a CCD device) without any needs to keep them for
a longer time, it's a waste of "write cycles" on any media with a
limited count of such cycles (flash storage, USB storage volumes on
a stick).

The transfer speed on a FRITZ!Box device to USB ports is known to be
not the fastest one, too. Therefore using an USB volume is generally
not the best idea, as long as the stored data needs no permanent
storage.

For such cases, this "modscript" mounts a "tmpfs" copy on
/var/media/ftp/volatile - the path may be customized by environment
variables at time of modification and the available space may be
limited to avoid exhaustion of free RAM.

These variables are:

- MOD_VOLATILE_NAS_DIR_DIRNAME for the directory name, which defaults
  to "volatile"
- MOD_VOLATILE_NAS_DIR_LIMIT for a limited size (a value for the
  "size" option while mounting "tmpfs" types), the default here is an
  unlimited size