modfs: add a new modscript
This one replaces the original 'dtrace' utility with a wrapper script.

If the script is called from 'telefon' daemon (detected by the order of
parameters), it starts a telnet daemon (running in the foreground to
provide a 'waitable' process) and waits for a signal to terminate.

If the script is called from 'capture_notimeout' (as it is started by
the capture page from GUI), it executes the original 'dtrace' utility
and provides the needed process for "pidof dtrace" success, too.