mod_show_vpn_on_overview: create a second copy of this script
to be used for FRITZ!OS versions up to (but not including) 07.24.

It seems easier to create a new version, which may be used up from
07.24 - first the removal of AVM's own datasource was changed and
secondly we'll start to use AVM's JS code to display our own line(s)
resulting in much lesser changes to AVM's original code and in much
lesser chances to get a conflict in future.